feat: allow renaming the libraries
Normally, the libraries built by GHDL are generated using the target
name as the library name.

However, this also holds for other tools as well - so if you want to
build different tools' targets from the same BUILD file, you must
have a way to untangle the target names, without changing the
library names.

This adds a `library_name` param, which achieves this. If you have
a library name conflict, modify one of the target names, and add
`library_name = "name that you want"`.